requirements.txt
requests==2.31.0
feedparser==6.0.11
pyyaml==6.0.1
